feat(chart): Old classes removed
