+++ src/lib/ui/app/Chart/drawing-tools/pane-view.ts
@@ -1,38 +1,64 @@
 import type {
   Coordinate,
+  IPrimitivePaneRenderer,
   IPrimitivePaneView,
   PrimitivePaneViewZOrder,
 } from '@santiment-network/chart-next'
-import type { RectanglePrimitive } from './rectangle/primitive.js'
+import type { DrawingPrimitive } from './primitive.js'
 
-import { RectangleAxisPaneRenderer } from './rectangle/renderer.js'
+import { DrawingAxisPaneRenderer } from './renderer.js'
 
-export abstract class DrawingToolAxisPaneView implements IPrimitivePaneView {
-  _source: RectanglePrimitive
-  _p1: number | null = null
-  _p2: number | null = null
-  _vertical: boolean = false
+export abstract class DrawingPaneView implements IPrimitivePaneView {
+  _source: DrawingPrimitive<any>
 
-  constructor(source: RectanglePrimitive, vertical: boolean) {
+  constructor(source: DrawingPrimitive<any>) {
     this._source = source
-    this._vertical = vertical
   }
 
-  abstract getPoints(): [Coordinate | null, Coordinate | null]
+  abstract update(): void
+
+  abstract renderer(): IPrimitivePaneRenderer | null
+}
+
+export abstract class DrawingAxisPaneView implements IPrimitivePaneView {
+  _source: DrawingPrimitive<any>
+  _points: (Coordinate | null)[] = []
+  _vertical: boolean = false
+
+  constructor(source: DrawingPrimitive<any>) {
+    this._source = source
+  }
+
+  abstract getPoints(): (Coordinate | null)[]
 
   update() {
-    ;[this._p1, this._p2] = this.getPoints()
+    this._points = this.getPoints()
   }
 
   renderer() {
-    return new RectangleAxisPaneRenderer(
-      this._p1,
-      this._p2,
+    return new DrawingAxisPaneRenderer(
+      this._points,
       this._source._options.fillColor,
       this._vertical,
     )
   }
   zOrder(): PrimitivePaneViewZOrder {
     return 'bottom'
+  }
+}
+
+export class DrawingPriceAxisPaneView extends DrawingAxisPaneView {
+  _vertical: boolean = true
+
+  getPoints(): (Coordinate | null)[] {
+    const series = this._source.series
+    return this._source.points.map((point) => series.priceToCoordinate(point.price))
+  }
+}
+
+export class DrawingTimeAxisPaneView extends DrawingAxisPaneView {
+  getPoints(): (Coordinate | null)[] {
+    const timeScale = this._source.chart.timeScale()
+    return this._source.points.map((point) => timeScale.timeToCoordinate(point.time))
   }
 }
