+++ src/lib/ui/app/Chart/drawing-tools/renderer.ts
